Regen aws_ec2 migration
